Remove train subfolder
