+++ ci/common/test-pyclaragenomics.sh
@@ -36,6 +36,7 @@ rsync -aP rsync://hgdownload-euro.soe.ucsc.edu/genome/admin/exe/linux.x86_64/big
 # Run tests.
 if [ "${TEST_ON_GPU}" == '1' ]; then
     ./example/run.sh
+    ./tests/run_v0.1.sh
     python -m pytest tests/
 else
     echo "No CPU tests."
