+++ tests/run_v0.1.sh
@@ -10,6 +10,7 @@
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
 
+# Example run script from version v0.1 to ensure backward compatibility.
 
 set -e
 
@@ -24,19 +25,21 @@ set -e
 # (These cover the first 10 Mb each of chr1, chr2, and chr3)
 
 # 5. example.sizes - lists the regions of the genome to cover (first 10 Mb each of chr1, chr2, and chr3)
+echo "Testing v0.1 run script to ensure backward compatibility"
 
 echo ""
 echo "Step 0: Initialize environment..."
 echo ""
-example_dir=$(readlink -f $(dirname "$0"))
+test_dir=$(readlink -f $(dirname "$0"))
 
+example_dir="$test_dir/../example"
 data_dir="$example_dir/data"
 ref_dir="$example_dir/reference"
 out_dir="$example_dir/result"
 
 root_dir=$(readlink -f "$example_dir/..")
 saved_model_dir="$root_dir/data/pretrained_models"
-config_dir="$root_dir/configs"
+
 # Switch to root directory before running script.
 cd $root_dir
 
@@ -109,8 +112,10 @@ echo ""
 python $root_dir/main.py --train \
     --train_files $out_dir/train_data.h5 \
     --val_files $out_dir/val_data.h5 \
-    --task both \
-    --bs 8 --reg_rounding 0 --cla_rounding 3 \
+    --model resnet --nblocks 5 --nfilt 15 --width 50 \
+    --dil 8 --task both --epochs 2 --afunc relu --mse_weight 0.001 \
+    --nblocks_cla 2 --nfilt_cla 15 --width_cla 50 --dil_cla 10 \
+    --pearson_weight 1 --bs 8 --reg_rounding 0 --cla_rounding 3 \
     --out_home $out_dir --label HSC.5M.model \
     --checkpoint_fname checkpoint.pth.tar \
     --save_freq=1 --eval_freq=1 --distributed
@@ -132,10 +137,6 @@ python $root_dir/calculate_baseline_metrics.py \
 echo ""
 echo "Step 6a: Run inference on test set with default peak calling setting..."
 echo ""
-
-# Feature alert: usage of --config_mparams. 
-##This option allows you to specify a custom model config file.
-
 # Note: change --weights_path to the path for your saved model!
 python $root_dir/main.py --infer \
     --infer_files $out_dir/test_data.h5 \
@@ -143,8 +144,10 @@ python $root_dir/main.py --infer \
     --sizes_file $ref_dir/hg19.auto.sizes \
     --infer_threshold 0.5 \
     --weights_path $out_dir/HSC.5M.model_latest/model_best.pth.tar \
-    --out_home $out_dir --label inference --config_mparams $config_dir/model_structure.yaml \
+    --out_home $out_dir --label inference \
     --result_fname HSC.5M.output --reg_rounding 0 --cla_rounding 3 \
+    --model resnet --nblocks 5 --nfilt 15 --width 50 --dil 8 \
+    --nblocks_cla 2 --nfilt_cla 15 --width_cla 50 --dil_cla 10 \
     --task both --num_workers 0 --gen_bigwig
 
 echo ""
@@ -189,6 +192,8 @@ python $root_dir/main.py --infer \
     --weights_path $out_dir/HSC.5M.model_latest/model_best.pth.tar \
     --out_home $out_dir --label inference --reg_rounding 0 --cla_rounding 3 \
     --result_fname HSC.5M.output.probs \
+    --model resnet --nblocks 5 --nfilt 15 --width 50 --dil 8 \
+    --nblocks_cla 2 --nfilt_cla 15 --width_cla 50 --dil_cla 10 \
     --task both --num_workers 0 --gen_bigwig
 
 macs2 bdgpeakcall -i $out_dir/inference_latest/test_data_HSC.5M.output.probs.peaks.bedGraph -o $out_dir/inference_latest/test_data_HSC.5M.output.peaks.narrowPeak -c 0.5
@@ -206,6 +211,8 @@ python $root_dir/main.py --infer \
     --weights_path $saved_model_dir/bulk_blood_data/5000000.7cell.resnet.5.2.15.8.50.0803.pth.tar \
     --out_home $out_dir --label inference.pretrained \
     --result_fname HSC.5M.output.pretrained --reg_rounding 0 --cla_rounding 3 \
+    --model resnet --nblocks 5 --nfilt 15 --width 50 --dil 8 \
+    --nblocks_cla 2 --nfilt_cla 15 --width_cla 50 --dil_cla 10 \
     --task both --num_workers 0 --gen_bigwig
 
 echo ""
